Backend/requirements.txt
flask
flask_socketio
supabase
python-dotenv